Extract angular configuration from ui/chrome (#34314)
