+++ src/legacy/core_plugins/tests_bundle/tests_entry_template.js
@@ -50,6 +50,7 @@ new CoreSystem({
     legacyMetadata: {
       version: '1.2.3',
       buildNum: 1234,
+      devMode: true,
       uiSettings: {
         defaults: ${JSON.stringify(defaultUiSettings, null, 2).split('\n').join('\n    ')},
         user: {}
