+++ src/legacy/ui/public/chrome/api/loading_count.js
@@ -19,8 +19,6 @@
 
 import * as Rx from 'rxjs';
 
-import { isSystemApiRequest } from '../../system_api';
-
 let newPlatformHttp;
 
 export function __newPlatformInit__(instance) {
@@ -30,27 +28,7 @@ export function __newPlatformInit__(instance) {
   newPlatformHttp = instance;
 }
 
-export function initLoadingCountApi(chrome, internals) {
-  /**
-   * Injected into angular module by ui/chrome angular integration
-   * and adds a root-level watcher that will capture the count of
-   * active $http requests on each digest loop and expose the count to
-   * the core.loadingCount api
-   * @param  {Angular.Scope} $rootScope
-   * @param  {HttpService} $http
-   * @return {undefined}
-   */
-  internals.capture$httpLoadingCount = function ($rootScope, $http) {
-    newPlatformHttp.addLoadingCount(new Rx.Observable(observer => {
-      const unwatch = $rootScope.$watch(() => {
-        const reqs = $http.pendingRequests || [];
-        observer.next(reqs.filter(req => !isSystemApiRequest(req)).length);
-      });
-
-      return unwatch;
-    }));
-  };
-
+export function initLoadingCountApi(chrome) {
   const manualCount$ = new Rx.BehaviorSubject(0);
   newPlatformHttp.addLoadingCount(manualCount$);
